base volume server

+++ storage/file_id.go
@@ -23,5 +23,5 @@ func ParseFileId(fid string) (*FileId, error) {
 	return nil, errors.New("Invalid fid " + fid)
 }
 func (n *FileId) String() string {
-	return fmt.Sprintf("%x-%x", n.VolumeId, n.Key)
+	return fmt.Sprintf("%x-%x", uint32(n.VolumeId), n.Key)
 }
